+++ demo/scripts/controls/sidePane/editorOptions/ExperimentalFeatures.tsx
@@ -7,13 +7,7 @@ export interface ExperimentalFeaturesProps {
     resetState: (callback: (state: BuildInPluginState) => void, resetEditor: boolean) => void;
 }
 
-const FeatureNames: { [key in ExperimentalFeatures]: string } = {
-    [ExperimentalFeatures.ListChain]: 'Enable List Chain for numbering list',
-    [ExperimentalFeatures.MergePastedLine]:
-        'When paste, try merge pasted content to the same line with existing content',
-    [ExperimentalFeatures.NewBullet]: 'Toggle bullet using VList',
-    [ExperimentalFeatures.NewIndentation]: 'Toggle indentation using VList',
-    [ExperimentalFeatures.NewNumbering]: 'Toggle numbering using VList',
+const FeatureNames: { [key in ExperimentalFeatures]?: string } = {
     [ExperimentalFeatures.SingleDirectionResize]: 'Resize an image horizontally or vertically',
     [ExperimentalFeatures.PasteWithLinkPreview]: 'Try retrieve link preview information when paste',
 };
