+++ packages/roosterjs-editor-api/lib/format/setIndentation.ts
@@ -1,13 +1,19 @@
-import experimentSetIndentation from '../experiment/experimentSetIndentation';
-import processList from '../utils/processList';
+import blockFormat from '../utils/blockFormat';
+import { BlockElement, IEditor, Indentation, RegionBase } from 'roosterjs-editor-types';
 import {
-    ChangeSource,
-    DocumentCommand,
-    IEditor,
-    Indentation,
-    QueryScope,
-    ExperimentalFeatures,
-} from 'roosterjs-editor-types';
+    collapseNodesInRegion,
+    createVListFromRegion,
+    findClosestElementAncestor,
+    getSelectedBlockElementsInRegion,
+    getTagOfNode,
+    isNodeInRegion,
+    splitBalancedNodeRange,
+    toArray,
+    unwrap,
+    wrap,
+} from 'roosterjs-editor-dom';
+
+const BlockWrapper = '<blockquote style="margin-top:0;margin-bottom:0"></blockquote>';
 
 /**
  * Set indentation at selection
@@ -18,32 +24,53 @@ import {
  * Indentation.Increase to increase indentation or Indentation.Decrease to decrease indentation
  */
 export default function setIndentation(editor: IEditor, indentation: Indentation) {
-    if (editor.isFeatureEnabled(ExperimentalFeatures.NewIndentation)) {
-        experimentSetIndentation(editor, indentation);
-    } else {
-        let command: DocumentCommand.Indent | DocumentCommand.Outdent =
-            indentation == Indentation.Increase ? DocumentCommand.Indent : DocumentCommand.Outdent;
-        editor.addUndoSnapshot(() => {
-            editor.focus();
-            let listNode = editor.getElementAtCursor('OL,UL');
-            let newNode: Node;
+    const handler = indentation == Indentation.Increase ? indent : outdent;
 
-            if (listNode) {
-                // There is already list node, setIndentation() will increase/decrease the list level,
-                // so we need to process the list when change indentation
-                newNode = processList(editor, command);
+    blockFormat(editor, (region, start, end) => {
+        const blocks = getSelectedBlockElementsInRegion(region, true /*createBlockIfEmpty*/);
+        const blockGroups: BlockElement[][] = [[]];
+
+        for (let i = 0; i < blocks.length; i++) {
+            const startNode = blocks[i].getStartNode();
+            const vList = createVListFromRegion(region, true /*includeSiblingLists*/, startNode);
+
+            if (vList) {
+                blockGroups.push([]);
+                while (blocks[i + 1] && vList.contains(blocks[i + 1].getStartNode())) {
+                    i++;
+                }
+                vList.setIndentation(start, end, indentation);
+                vList.writeBack();
             } else {
-                // No existing list node, browser will create <Blockquote> node for indentation.
-                // We need to set top and bottom margin to 0 to avoid unnecessary spaces
-                editor.getDocument().execCommand(command, false, null);
-                editor.queryElements('BLOCKQUOTE', QueryScope.OnSelection, node => {
-                    newNode = newNode || node;
-                    node.style.marginTop = '0px';
-                    node.style.marginBottom = '0px';
-                });
+                blockGroups[blockGroups.length - 1].push(blocks[i]);
+            }
+        }
+
+        blockGroups.forEach(group => handler(region, group));
+    });
+}
+
+function indent(region: RegionBase, blocks: BlockElement[]) {
+    const nodes = collapseNodesInRegion(region, blocks);
+    wrap(nodes, BlockWrapper);
+}
+
+function outdent(region: RegionBase, blocks: BlockElement[]) {
+    blocks.forEach(blockElement => {
+        let node = blockElement.collapseToSingleElement();
+        const quote = findClosestElementAncestor(node, region.rootNode, 'blockquote');
+        if (quote) {
+            if (node == quote) {
+                node = wrap(toArray(node.childNodes));
             }
 
-            return newNode;
-        }, ChangeSource.Format);
-    }
+            while (isNodeInRegion(region, node) && getTagOfNode(node) != 'BLOCKQUOTE') {
+                node = splitBalancedNodeRange(node);
+            }
+
+            if (isNodeInRegion(region, node)) {
+                unwrap(node);
+            }
+        }
+    });
 }
