+++ packages/roosterjs-editor-plugins/lib/plugins/ContentEdit/features/listFeatures.ts
@@ -1,4 +1,3 @@
-import experimentToggleListType from 'roosterjs-editor-api/lib/experiment/experimentToggleListType';
 import {
     experimentCommitListChains,
     setIndentation,
@@ -22,10 +21,8 @@ import {
     NodeType,
     PluginKeyboardEvent,
     PositionType,
-    ListType,
     QueryScope,
     RegionBase,
-    ExperimentalFeatures,
 } from 'roosterjs-editor-types';
 
 /**
@@ -160,14 +157,10 @@ const AutoBullet: BuildInEditFeature<PluginKeyboardEvent> = {
                     } else if (textBeforeCursor.indexOf('1.') == 0) {
                         prepareAutoBullet(editor, rangeToDelete);
                         toggleNumbering(editor);
-                    } else if (
-                        editor.isFeatureEnabled(ExperimentalFeatures.ListChain) &&
-                        (regions = editor.getSelectedRegions()) &&
-                        regions.length == 1
-                    ) {
+                    } else if ((regions = editor.getSelectedRegions()) && regions.length == 1) {
                         const num = parseInt(textBeforeCursor);
                         prepareAutoBullet(editor, rangeToDelete);
-                        experimentToggleListType(editor, ListType.Ordered, num);
+                        toggleNumbering(editor, num);
                     }
                 },
                 null /*changeSource*/,
@@ -195,9 +188,7 @@ const MaintainListChain: BuildInEditFeature<PluginKeyboardEvent> = {
 };
 
 function getListChains(editor: IEditor) {
-    return editor.isFeatureEnabled(ExperimentalFeatures.ListChain)
-        ? VListChain.createListChains(editor.getSelectedRegions())
-        : [];
+    return VListChain.createListChains(editor.getSelectedRegions());
 }
 
 function prepareAutoBullet(editor: IEditor, range: Range) {
