+++ packages/roosterjs-editor-plugins/lib/plugins/CutPasteListChain/CutPasteListChain.ts
@@ -6,7 +6,6 @@ import {
     IEditor,
     PluginEvent,
     PluginEventType,
-    ExperimentalFeatures,
 } from 'roosterjs-editor-types';
 
 /**
@@ -74,9 +73,7 @@ export default class CutPasteListChain implements EditorPlugin {
     };
 
     private cacheListChains(source: ChangeSource) {
-        if (this.editor.isFeatureEnabled(ExperimentalFeatures.ListChain)) {
-            this.chains = VListChain.createListChains(this.editor.getSelectedRegions());
-            this.expectedChangeSource = source;
-        }
+        this.chains = VListChain.createListChains(this.editor.getSelectedRegions());
+        this.expectedChangeSource = source;
     }
 }
